requirements.txt
Django==1.11
gunicorn==19.7.1
whitenoise==3.3.1
djangorestframework==2.0